requirements.txt
flask
Flask-MySQLdb
argon2-cffi